refactor: 개인정보 입력 화면 UI State로 리팩터링
